refactor: move IEmailData to interfaces folder
