--- src/types/index.ts
@@ -1,6 +1,0 @@
-export interface IEmailData {
-  from: string;
-  to: string;
-  subject: string;
-  html: string;
-}
